pyproject.toml
[tool.poetry]
name = "lut-convert"
version = "0.0.0"
description = "LUT converter"
authors = ["Christian Wygoda <christian.wygoda@wygoda.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
numpy = "^1.26.4"
click = "^8.1.7"

[tool.poetry.scripts]
lut-convert = "lut_convert.cli:cli"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true

[tool.ruff]
line-length = 88

[tool.ruff.lint]
extend-ignore = ["E501", "UP007", "UP034"]
select = [
    "C9",
    "E",
    "F",
    "I",
    "W",
    "UP"
]

[tool.ruff.lint.mccabe]
max-complexity = 8

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
